interface/requirements.txt
beautifulsoup4==4.9.3
boto3==1.17.78
botocore==1.20.78
certifi==2020.12.5
IMAPClient==2.2.0
install==1.3.4
jmespath==0.10.0
markdownify==0.8.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-slugify==5.0.2
redis==3.5.3
s3transfer==0.4.2
six==1.16.0
soupsieve==2.2.1
text-unidecode==1.3
urllib3==1.26.4
